refactor(core): parse eval frontmatter with gray-matter
Replace the hand-rolled YAML-ish frontmatter parser in eval-metadata.ts
with the gray-matter package, matching how supabase/supabase parses
markdown frontmatter in the docs app. This drops ~165 lines of bespoke
tokenizing in favor of a maintained YAML parser, while preserving the
existing behavior: required frontmatter, product/products + topic/topics
aliases, token normalization (lowercase + whitespace collapse), and the
same error messages.

Verified byte-for-byte identical metadata and body output across all 23
PROMPT.md files before/after the swap.

The Node-only parser lives in a new eval-markdown.ts so eval-metadata.ts
stays browser-safe: the web app imports schemas from it, and keeping
gray-matter (CJS, uses eval/Buffer) out of that module shrinks the web
bundle from 481KB to 383KB and clears a Rollup eval warning.
