+++ packages/core/package.json
@@ -7,7 +7,8 @@
   "types": "src/index.ts",
   "exports": {
     ".": "./src/index.ts",
-    "./eval-metadata": "./src/eval-metadata.ts"
+    "./eval-metadata": "./src/eval-metadata.ts",
+    "./eval-markdown": "./src/eval-markdown.ts"
   },
   "dependencies": {
     "@ai-sdk/mcp": "catalog:",
@@ -16,6 +17,7 @@
     "@supabase/supabase-js": "catalog:",
     "ai": "catalog:",
     "executor": "1.4.29",
+    "gray-matter": "^4.0.3",
     "typescript": "catalog:",
     "zod": "catalog:"
   }
